Updated Cart and MusicSlippers
also changed cost to a double as a cost can have a decimal amount

+++ src/Product.java
@@ -1,7 +1,7 @@
 public interface Product{
 	public String getName();
 	public String getDescription();
-	public int getCost();
+	public double getCost();
 	public void setName(String name);
 	public void setDescription(String description);
 	public void setCost(int cost);
